requirements.txt
torch==2.3.1
transformers==4.41.2
faiss-gpu==1.7.2
opencv-python==4.10
bitsandbytes==0.43.1
safetensors==0.4.3
sentence-transformers==3.0.1
streamlit==1.35.0
